integers changed to uint32

+++ internal/filesystem/filesystem.go
@@ -16,9 +16,9 @@ type FileSystem struct {
 	currentDirectory directory.Directory
 }
 
-func FormatFilesystem(sizeInBytes int, blockSize int) (*FileSystem, error) {
+func FormatFilesystem(sizeInBytes uint32, blockSize uint32) (*FileSystem, error) {
 	fileSystem := FileSystem{}
-	fileSystem.Superblock = superblock.NewSuperblock(uint32(sizeInBytes), uint32(blockSize))
+	fileSystem.Superblock = superblock.NewSuperblock(sizeInBytes, blockSize)
 	fileSystem.BlockBitmap = bitmap.NewBitmap(fileSystem.Superblock.BlockCount)
 	fileSystem.InodeBitmap = bitmap.NewBitmap(fileSystem.Superblock.InodeCount)
 
@@ -33,14 +33,14 @@ func FormatFilesystem(sizeInBytes int, blockSize int) (*FileSystem, error) {
 	fileSystem.InodeBitmap.WriteAt(fileSystem.dataFile, fileSystem.GetInodeBitmapOffset())
 
 	inodeTableSize := fileSystem.Superblock.InodeCount * fileSystem.Superblock.InodeSize
-	fileSystem.ReserveSpaceInFile(fileSystem.GetInodeTableOffset(), inodeTableSize+uint32(sizeInBytes))
+	fileSystem.ReserveSpaceInFile(fileSystem.GetInodeTableOffset(), inodeTableSize+sizeInBytes)
 
 	fileSystem.CreateRootDirectory()
 
 	return &fileSystem, nil
 }
 
-func (fs FileSystem) ReserveSpaceInFile(offset int, size uint32) error {
+func (fs FileSystem) ReserveSpaceInFile(offset uint32, size uint32) error {
 	data := make([]byte, size)
 
 	_, err := fs.dataFile.WriteAt(data, int64(offset))
@@ -85,7 +85,7 @@ func (fs FileSystem) CreateFirstFile(name string) error {
 	fs.Superblock.FreeInodeCount--
 
 	fileInode := inode.NewInode(true, 777, 0, 0, []uint32{blockIndex})
-	inodeOffset := fs.GetInodeTableOffset() + int(fs.Superblock.InodeSize*inodeIndex)
+	inodeOffset := fs.GetInodeTableOffset() + fs.Superblock.InodeSize*inodeIndex
 	fileInode.WriteAt(fs.dataFile, inodeOffset)
 
 	fs.currentDirectory.AddFile(inodeIndex, name)
@@ -93,20 +93,20 @@ func (fs FileSystem) CreateFirstFile(name string) error {
 	return nil
 }
 
-func (fs FileSystem) GetBlockBitmapOffset() int {
+func (fs FileSystem) GetBlockBitmapOffset() uint32 {
 	return fs.Superblock.Size()
 }
 
-func (fs FileSystem) GetInodeBitmapOffset() int {
-	return fs.GetBlockBitmapOffset() + int(fs.BlockBitmap.Size)
+func (fs FileSystem) GetInodeBitmapOffset() uint32 {
+	return fs.GetBlockBitmapOffset() + fs.BlockBitmap.Size
 }
 
-func (fs FileSystem) GetInodeTableOffset() int {
-	return fs.GetInodeBitmapOffset() + int(fs.InodeBitmap.Size)
+func (fs FileSystem) GetInodeTableOffset() uint32 {
+	return fs.GetInodeBitmapOffset() + fs.InodeBitmap.Size
 }
 
-func (fs FileSystem) GetDataBlocksOffset() int {
-	return fs.GetInodeTableOffset() + int(fs.Superblock.InodeCount*fs.Superblock.InodeSize)
+func (fs FileSystem) GetDataBlocksOffset() uint32 {
+	return fs.GetInodeTableOffset() + fs.Superblock.InodeCount*fs.Superblock.InodeSize
 }
 
 func (fs FileSystem) CloseDataFile() error {
